% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GimmeMSD}
\alias{GimmeMSD}
\title{GimmeMSD}
\usage{
GimmeMSD(y, digits = 3, na.rm = TRUE, se = FALSE)
}
\arguments{
\item{y}{the numerical vector.}

\item{digits}{the number of digits.}

\item{na.rm}{if they want to remove NAs from the computation}

\item{se}{if TRUE it will use the standard error instead of
the standard deviation}
}
\value{
a string.
}
\description{
A function that computes means and standard deviations (or standard error)
of a numeric vector and returns a string
}
\examples{
GimmeMSD(c(1, 4, 8, 2, 3, 9, 10))
}
