% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GimmeResult}
\alias{GimmeResult}
\title{GimmeResult}
\usage{
GimmeResult(test, digits = c(3, 2))
}
\arguments{
\item{test}{is the object resulting from a test.}

\item{digits}{a vector containing the number of required digits for the
degrees of freedom (second element in the vector), statistics and p-value
(first element of the vector). Default = c(3, 2)}
}
\value{
a string.
}
\description{
A function that takes the results of a t.test or other common tests and
returns a string formatted similarly to APA style
}
\examples{
result <- t.test(1:10, 11:20)
GimmeResult(result)
}
