% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GimmeResult2}
\alias{GimmeResult2}
\title{GimmeResult2}
\usage{
GimmeResult2(test, digits = c(3, 2))
}
\arguments{
\item{test}{is the object resulting from a test.}

\item{digits}{a vector containing the number of required digits for the
degrees of freedom (second element in the vector), statistics and p-value
(first element of the vector). Default = c(3, 2)}
}
\value{
a string.
}
\description{
A function that takes the results of a chi squared test and
returns a string formatted similarly to APA style
}
