% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GimmeStar}
\alias{GimmeStar}
\title{GimmeStar}
\usage{
GimmeStar(p, not_trend = FALSE)
}
\arguments{
\item{p}{the p-value.}

\item{not_trend}{if FALSE returns "+" for p-values between 0.05 and 0.1}
}
\value{
a string.
}
\description{
A function that reads the p-values and returns the asterisks
}
\examples{
GimmeStar(0.01) ## returns **
GimmeStar(0.010)
GimmeStar(0.07)
GimmeStar(0.07, not_trend = TRUE)
sapply(c(0.1, 0.04, 0.64, 0.02), GimmeStar) ## \"+\" \"*\" \"\"  \"*\"
}
