% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GimmeStarCor}
\alias{GimmeStarCor}
\title{GimmeStarCor}
\usage{
GimmeStarCor(r, not_trend = FALSE)
}
\arguments{
\item{r}{the correlation value.}

\item{not_trend}{if FALSE returns "+" for rhos between 0.3 and 0.1}
}
\value{
a string.
}
\description{
A function that reads correlations and returns the asterisks
}
\examples{
GimmeStarCor(0.5) ## returns **
GimmeStarCor(0.1)
GimmeStarCor(0.2)
GimmeStarCor(0.2, not_trend = TRUE)
sapply(c(0.1, 0.04, 0.64, 0.02), GimmeStarCor)
}
