% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{l.}
\alias{l.}
\title{Log-transform a numeric variable}
\usage{
l.(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector containing the natural logarithm of \code{x}.
}
\description{
Applies the natural logarithm to a numeric vector.
}
\details{
This is a convenience wrapper around \code{\link[base]{log}}. Values must
satisfy the usual requirements of the logarithm: zero values return
\code{-Inf}, while negative values produce \code{NaN}.
}
\examples{
x <- c(1, 2, 5, 10)
l.(x)

}
\seealso{
\code{\link[base]{log}}
}
