% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{my_F_to_eta2}
\alias{my_F_to_eta2}
\title{Convert an F statistic to partial eta squared}
\usage{
my_F_to_eta2(f, df, df_error)
}
\arguments{
\item{f}{A numeric value or vector containing F statistics.}

\item{df}{A numeric value or vector containing the numerator degrees of
freedom associated with each F statistic.}

\item{df_error}{A numeric value or vector containing the denominator,
or error, degrees of freedom associated with each F statistic.}
}
\value{
A numeric value or vector containing partial eta squared
(\eqn{\eta_p^2}).
}
\description{
Converts an observed F statistic and its degrees of freedom to partial
eta squared.
}
\details{
Partial eta squared is calculated from an F statistic as:

\deqn{
\eta_p^2 =
\frac{F \times df}
     {F \times df + df_{error}}
}

The function is vectorized, so multiple F statistics can be converted in
a single call.
}
\examples{
# F = 5.4, numerator df = 2, denominator df = 97
my_F_to_eta2(
  f = 5.4,
  df = 2,
  df_error = 97
)

# Convert several F statistics
my_F_to_eta2(
  f = c(2.1, 5.4, 10.2),
  df = c(1, 2, 1),
  df_error = c(98, 97, 98)
)

}
