% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{strInNum}
\alias{strInNum}
\title{strInNum.}
\usage{
strInNum(input)
}
\arguments{
\item{input}{is the factor.}
}
\value{
the numeric version of the factor.
}
\description{
Function that takes a factor that is numeric into a number.
The factors should only contain digits.
}
\examples{
strInNum(as.factor(c("1", "2", "3"))) ## returns 1 2 3
}
