% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{z.}
\alias{z.}
\title{Standardize a numeric variable}
\usage{
z.(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
A standardized object with the same dimensions as \code{x}. For a
vector, the returned object is a one-column matrix because the function
relies on \code{\link[base]{scale}}.
}
\description{
Standardizes a numeric vector or matrix by subtracting the mean and
dividing by the standard deviation.
}
\details{
This is a convenience wrapper around \code{\link[base]{scale}} using its
default arguments. For a numeric variable \eqn{x}, the transformation is

\deqn{z = (x - \bar{x}) / s}

where \eqn{\bar{x}} is the sample mean and \eqn{s} is the sample standard
deviation.
}
\examples{
x <- c(10, 12, 14, 16)
z.(x)

mean(z.(x))
sd(z.(x))

}
\seealso{
\code{\link[base]{scale}}, \code{\link{c.}}
}
